More directory restructuring.
