+++ examples/ble_nus/main.cpp
@@ -55,7 +55,7 @@ int main(int argc, char* argv[]) {
         auto characteristic_rx = device->get_characteristic("6e400001-b5a3-f393-e0a9-e50e24dcca9e",
                                                             "6e400002-b5a3-f393-e0a9-e50e24dcca9e");
         if (characteristic_tx != nullptr) {
-            characteristic_tx->ValueChanged = [&](std::vector<uint8_t>& new_value) { /* Data! */ };
+            characteristic_tx->ValueChanged = [&](std::vector<uint8_t> new_value) { /* Data! */ };
             characteristic_tx->StartNotify();
         }
