+++ src/simpledbus/common/ObjectManager.cpp
