+++ src/simpledbus/common/ObjectManager.h
