+++ src/simpledbus/common/Properties.cpp
