+++ src/simpledbus/common/Properties.h
